(test) Prune transition frontier
